Refactored the API tests into their own folder so Lab can run client tests too
