+++ package.json
@@ -53,7 +53,6 @@
     "http-proxy": "^1.13.1",
     "http-server": "^0.8.5",
     "lab": "^8.2.0",
-    "mocha": "^2.3.4",
     "sequelize-fixtures": "^0.4.8",
     "shot": "^3.0.0",
     "sinon": "^1.17.3"
