+++ test/api/construct.js
@@ -4,12 +4,10 @@
 
 const Fixtures = require('sequelize-fixtures');
 
-const models = require('../models');
+const models = require('../../models');
 
 
 exports.fixtures = function(fixtures, done) {
-  const options = {cascade: true};
-
   models.sequelize.sync({force: true}).then(() => {
     Fixtures.loadFile(fixtures, models).then(() => {
       if (done) {
